LayoutTests/svg/custom/global-constructors-expected.txt

Test to make sure we expose all the global constructor objects requested by http://www.w3.org/TR/SVG11/ecmascript-binding.html

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS SVGException.toString() is '[object SVGExceptionConstructor]'
PASS SVGElement.toString() is '[object SVGElementConstructor]'
PASS SVGAnimatedBoolean.toString() is '[object SVGAnimatedBooleanConstructor]'
PASS SVGAnimatedString.toString() is '[object SVGAnimatedStringConstructor]'
PASS SVGStringList.toString() is '[object SVGStringListConstructor]'
PASS SVGAnimatedEnumeration.toString() is '[object SVGAnimatedEnumerationConstructor]'
PASS SVGAnimatedInteger.toString() is '[object SVGAnimatedIntegerConstructor]'
PASS SVGNumber.toString() is '[object SVGNumberConstructor]'
PASS SVGAnimatedNumber.toString() is '[object SVGAnimatedNumberConstructor]'
PASS SVGNumberList.toString() is '[object SVGNumberListConstructor]'
PASS SVGAnimatedNumberList.toString() is '[object SVGAnimatedNumberListConstructor]'
PASS SVGLength.toString() is '[object SVGLengthConstructor]'
PASS SVGAnimatedLength.toString() is '[object SVGAnimatedLengthConstructor]'
PASS SVGLengthList.toString() is '[object SVGLengthListConstructor]'
PASS SVGAnimatedLengthList.toString() is '[object SVGAnimatedLengthListConstructor]'
PASS SVGAngle.toString() is '[object SVGAngleConstructor]'
PASS SVGAnimatedAngle.toString() is '[object SVGAnimatedAngleConstructor]'
PASS SVGColor.toString() is '[object SVGColorConstructor]'
FAIL SVGICCColor.toString() should be [object SVGICCColorConstructor]. Threw exception ReferenceError: Can't find variable: SVGICCColor
PASS SVGRect.toString() is '[object SVGRectConstructor]'
PASS SVGAnimatedRect.toString() is '[object SVGAnimatedRectConstructor]'
FAIL SVGStylable.toString() should be [object SVGStylableConstructor]. Threw exception ReferenceError: Can't find variable: SVGStylable
FAIL SVGLocatable.toString() should be [object SVGLocatableConstructor]. Threw exception ReferenceError: Can't find variable: SVGLocatable
FAIL SVGTransformable.toString() should be [object SVGTransformableConstructor]. Threw exception ReferenceError: Can't find variable: SVGTransformable
FAIL SVGTests.toString() should be [object SVGTestsConstructor]. Threw exception ReferenceError: Can't find variable: SVGTests
FAIL SVGLangSpace.toString() should be [object SVGLangSpaceConstructor]. Threw exception ReferenceError: Can't find variable: SVGLangSpace
FAIL SVGExternalResourcesRequired.toString() should be [object SVGExternalResourcesRequiredConstructor]. Threw exception ReferenceError: Can't find variable: SVGExternalResourcesRequired
FAIL SVGFitToViewBox.toString() should be [object SVGFitToViewBoxConstructor]. Threw exception ReferenceError: Can't find variable: SVGFitToViewBox
FAIL SVGZoomAndPan.toString() should be [object SVGZoomAndPanConstructor]. Threw exception ReferenceError: Can't find variable: SVGZoomAndPan
FAIL SVGViewSpec.toString() should be [object SVGViewSpecConstructor]. Threw exception ReferenceError: Can't find variable: SVGViewSpec
FAIL SVGURIReference.toString() should be [object SVGURIReferenceConstructor]. Threw exception ReferenceError: Can't find variable: SVGURIReference
FAIL SVGCSSRule.toString() should be [object SVGCSSRuleConstructor]. Threw exception ReferenceError: Can't find variable: SVGCSSRule
PASS SVGDocument.toString() is '[object SVGDocumentConstructor]'
PASS SVGSVGElement.toString() is '[object SVGSVGElementConstructor]'
PASS SVGGElement.toString() is '[object SVGGElementConstructor]'
PASS SVGDefsElement.toString() is '[object SVGDefsElementConstructor]'
PASS SVGDescElement.toString() is '[object SVGDescElementConstructor]'
PASS SVGTitleElement.toString() is '[object SVGTitleElementConstructor]'
PASS SVGSymbolElement.toString() is '[object SVGSymbolElementConstructor]'
PASS SVGUseElement.toString() is '[object SVGUseElementConstructor]'
PASS SVGElementInstance.toString() is '[object SVGElementInstanceConstructor]'
PASS SVGElementInstanceList.toString() is '[object SVGElementInstanceListConstructor]'
PASS SVGImageElement.toString() is '[object SVGImageElementConstructor]'
PASS SVGSwitchElement.toString() is '[object SVGSwitchElementConstructor]'
PASS SVGStyleElement.toString() is '[object SVGStyleElementConstructor]'
PASS SVGPoint.toString() is '[object SVGPointConstructor]'
PASS SVGPointList.toString() is '[object SVGPointListConstructor]'
PASS SVGMatrix.toString() is '[object SVGMatrixConstructor]'
PASS SVGTransform.toString() is '[object SVGTransformConstructor]'
PASS SVGTransformList.toString() is '[object SVGTransformListConstructor]'
PASS SVGAnimatedTransformList.toString() is '[object SVGAnimatedTransformListConstructor]'
PASS SVGPreserveAspectRatio.toString() is '[object SVGPreserveAspectRatioConstructor]'
PASS SVGAnimatedPreserveAspectRatio.toString() is '[object SVGAnimatedPreserveAspectRatioConstructor]'
PASS SVGPathSeg.toString() is '[object SVGPathSegConstructor]'
PASS SVGPathSegClosePath.toString() is '[object SVGPathSegClosePathConstructor]'
PASS SVGPathSegMovetoAbs.toString() is '[object SVGPathSegMovetoAbsConstructor]'
PASS SVGPathSegMovetoRel.toString() is '[object SVGPathSegMovetoRelConstructor]'
PASS SVGPathSegLinetoAbs.toString() is '[object SVGPathSegLinetoAbsConstructor]'
PASS SVGPathSegLinetoRel.toString() is '[object SVGPathSegLinetoRelConstructor]'
PASS SVGPathSegCurvetoCubicAbs.toString() is '[object SVGPathSegCurvetoCubicAbsConstructor]'
PASS SVGPathSegCurvetoCubicRel.toString() is '[object SVGPathSegCurvetoCubicRelConstructor]'
PASS SVGPathSegCurvetoQuadraticAbs.toString() is '[object SVGPathSegCurvetoQuadraticAbsConstructor]'
PASS SVGPathSegCurvetoQuadraticRel.toString() is '[object SVGPathSegCurvetoQuadraticRelConstructor]'
PASS SVGPathSegArcAbs.toString() is '[object SVGPathSegArcAbsConstructor]'
PASS SVGPathSegArcRel.toString() is '[object SVGPathSegArcRelConstructor]'
PASS SVGPathSegLinetoHorizontalAbs.toString() is '[object SVGPathSegLinetoHorizontalAbsConstructor]'
PASS SVGPathSegLinetoHorizontalRel.toString() is '[object SVGPathSegLinetoHorizontalRelConstructor]'
PASS SVGPathSegLinetoVerticalAbs.toString() is '[object SVGPathSegLinetoVerticalAbsConstructor]'
PASS SVGPathSegLinetoVerticalRel.toString() is '[object SVGPathSegLinetoVerticalRelConstructor]'
PASS SVGPathSegCurvetoCubicSmoothAbs.toString() is '[object SVGPathSegCurvetoCubicSmoothAbsConstructor]'
PASS SVGPathSegCurvetoCubicSmoothRel.toString() is '[object SVGPathSegCurvetoCubicSmoothRelConstructor]'
PASS SVGPathSegCurvetoQuadraticSmoothAbs.toString() is '[object SVGPathSegCurvetoQuadraticSmoothAbsConstructor]'
PASS SVGPathSegCurvetoQuadraticSmoothRel.toString() is '[object SVGPathSegCurvetoQuadraticSmoothRelConstructor]'
PASS SVGPathSegList.toString() is '[object SVGPathSegListConstructor]'
FAIL SVGAnimatedPathData.toString() should be [object SVGAnimatedPathDataConstructor]. Threw exception ReferenceError: Can't find variable: SVGAnimatedPathData
PASS SVGPathElement.toString() is '[object SVGPathElementConstructor]'
PASS SVGRectElement.toString() is '[object SVGRectElementConstructor]'
PASS SVGCircleElement.toString() is '[object SVGCircleElementConstructor]'
PASS SVGEllipseElement.toString() is '[object SVGEllipseElementConstructor]'
PASS SVGLineElement.toString() is '[object SVGLineElementConstructor]'
FAIL SVGAnimatedPoints.toString() should be [object SVGAnimatedPointsConstructor]. Threw exception ReferenceError: Can't find variable: SVGAnimatedPoints
PASS SVGPolylineElement.toString() is '[object SVGPolylineElementConstructor]'
PASS SVGPolygonElement.toString() is '[object SVGPolygonElementConstructor]'
PASS SVGTextContentElement.toString() is '[object SVGTextContentElementConstructor]'
PASS SVGTextPositioningElement.toString() is '[object SVGTextPositioningElementConstructor]'
PASS SVGTextElement.toString() is '[object SVGTextElementConstructor]'
PASS SVGTSpanElement.toString() is '[object SVGTSpanElementConstructor]'
PASS SVGTRefElement.toString() is '[object SVGTRefElementConstructor]'
PASS SVGTextPathElement.toString() is '[object SVGTextPathElementConstructor]'
PASS SVGAltGlyphElement.toString() is '[object SVGAltGlyphElementConstructor]'
PASS SVGAltGlyphDefElement.toString() is '[object SVGAltGlyphDefElementConstructor]'
PASS SVGAltGlyphItemElement.toString() is '[object SVGAltGlyphItemElementConstructor]'
PASS SVGGlyphRefElement.toString() is '[object SVGGlyphRefElementConstructor]'
PASS SVGPaint.toString() is '[object SVGPaintConstructor]'
PASS SVGMarkerElement.toString() is '[object SVGMarkerElementConstructor]'
FAIL SVGColorProfileElement.toString() should be [object SVGColorProfileElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGColorProfileElement
FAIL SVGColorProfileRule.toString() should be [object SVGColorProfileRuleConstructor]. Threw exception ReferenceError: Can't find variable: SVGColorProfileRule
PASS SVGGradientElement.toString() is '[object SVGGradientElementConstructor]'
PASS SVGLinearGradientElement.toString() is '[object SVGLinearGradientElementConstructor]'
PASS SVGRadialGradientElement.toString() is '[object SVGRadialGradientElementConstructor]'
PASS SVGStopElement.toString() is '[object SVGStopElementConstructor]'
PASS SVGPatternElement.toString() is '[object SVGPatternElementConstructor]'
PASS SVGClipPathElement.toString() is '[object SVGClipPathElementConstructor]'
PASS SVGMaskElement.toString() is '[object SVGMaskElementConstructor]'
PASS SVGFilterElement.toString() is '[object SVGFilterElementConstructor]'
FAIL SVGFilterPrimitiveStandardAttributes.toString() should be [object SVGFilterPrimitiveStandardAttributesConstructor]. Threw exception ReferenceError: Can't find variable: SVGFilterPrimitiveStandardAttributes
PASS SVGFEBlendElement.toString() is '[object SVGFEBlendElementConstructor]'
PASS SVGFEColorMatrixElement.toString() is '[object SVGFEColorMatrixElementConstructor]'
PASS SVGFEComponentTransferElement.toString() is '[object SVGFEComponentTransferElementConstructor]'
PASS SVGComponentTransferFunctionElement.toString() is '[object SVGComponentTransferFunctionElementConstructor]'
PASS SVGFEFuncRElement.toString() is '[object SVGFEFuncRElementConstructor]'
PASS SVGFEFuncGElement.toString() is '[object SVGFEFuncGElementConstructor]'
PASS SVGFEFuncBElement.toString() is '[object SVGFEFuncBElementConstructor]'
PASS SVGFEFuncAElement.toString() is '[object SVGFEFuncAElementConstructor]'
PASS SVGFECompositeElement.toString() is '[object SVGFECompositeElementConstructor]'
PASS SVGFEConvolveMatrixElement.toString() is '[object SVGFEConvolveMatrixElementConstructor]'
PASS SVGFEDiffuseLightingElement.toString() is '[object SVGFEDiffuseLightingElementConstructor]'
PASS SVGFEDistantLightElement.toString() is '[object SVGFEDistantLightElementConstructor]'
PASS SVGFEPointLightElement.toString() is '[object SVGFEPointLightElementConstructor]'
PASS SVGFESpotLightElement.toString() is '[object SVGFESpotLightElementConstructor]'
PASS SVGFEDisplacementMapElement.toString() is '[object SVGFEDisplacementMapElementConstructor]'
PASS SVGFEFloodElement.toString() is '[object SVGFEFloodElementConstructor]'
PASS SVGFEGaussianBlurElement.toString() is '[object SVGFEGaussianBlurElementConstructor]'
PASS SVGFEImageElement.toString() is '[object SVGFEImageElementConstructor]'
PASS SVGFEMergeElement.toString() is '[object SVGFEMergeElementConstructor]'
PASS SVGFEMergeNodeElement.toString() is '[object SVGFEMergeNodeElementConstructor]'
PASS SVGFEMorphologyElement.toString() is '[object SVGFEMorphologyElementConstructor]'
PASS SVGFEOffsetElement.toString() is '[object SVGFEOffsetElementConstructor]'
PASS SVGFESpecularLightingElement.toString() is '[object SVGFESpecularLightingElementConstructor]'
PASS SVGFETileElement.toString() is '[object SVGFETileElementConstructor]'
PASS SVGFETurbulenceElement.toString() is '[object SVGFETurbulenceElementConstructor]'
PASS SVGCursorElement.toString() is '[object SVGCursorElementConstructor]'
PASS SVGAElement.toString() is '[object SVGAElementConstructor]'
PASS SVGViewElement.toString() is '[object SVGViewElementConstructor]'
PASS SVGScriptElement.toString() is '[object SVGScriptElementConstructor]'
FAIL SVGEvent.toString() should be [object SVGEventConstructor]. Threw exception ReferenceError: Can't find variable: SVGEvent
PASS SVGZoomEvent.toString() is '[object SVGZoomEventConstructor]'
FAIL SVGAnimationElement.toString() should be [object SVGAnimationElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGAnimationElement
PASS SVGAnimateElement.toString() is '[object SVGAnimateElementConstructor]'
PASS SVGSetElement.toString() is '[object SVGSetElementConstructor]'
PASS SVGAnimateMotionElement.toString() is '[object SVGAnimateMotionElementConstructor]'
PASS SVGMPathElement.toString() is '[object SVGMPathElementConstructor]'
PASS SVGAnimateColorElement.toString() is '[object SVGAnimateColorElementConstructor]'
PASS SVGAnimateTransformElement.toString() is '[object SVGAnimateTransformElementConstructor]'
PASS SVGFontElement.toString() is '[object SVGFontElementConstructor]'
PASS SVGGlyphElement.toString() is '[object SVGGlyphElementConstructor]'
PASS SVGMissingGlyphElement.toString() is '[object SVGMissingGlyphElementConstructor]'
PASS SVGHKernElement.toString() is '[object SVGHKernElementConstructor]'
PASS SVGVKernElement.toString() is '[object SVGVKernElementConstructor]'
PASS SVGFontFaceElement.toString() is '[object SVGFontFaceElementConstructor]'
PASS SVGFontFaceSrcElement.toString() is '[object SVGFontFaceSrcElementConstructor]'
PASS SVGFontFaceUriElement.toString() is '[object SVGFontFaceUriElementConstructor]'
PASS SVGFontFaceFormatElement.toString() is '[object SVGFontFaceFormatElementConstructor]'
PASS SVGFontFaceNameElement.toString() is '[object SVGFontFaceNameElementConstructor]'
FAIL SVGDefinitionSrcElement.toString() should be [object SVGDefinitionSrcElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGDefinitionSrcElement
PASS SVGMetadataElement.toString() is '[object SVGMetadataElementConstructor]'
PASS SVGForeignObjectElement.toString() is '[object SVGForeignObjectElementConstructor]'
PASS successfullyParsed is true

TEST COMPLETE
